refactor: extract to a custom hook
