+++ src/components/Button/Button.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/button-has-type */
-import React, { AriaAttributes, forwardRef, useCallback, useEffect, useMemo, useRef, useState } from "react";
+import React, { AriaAttributes, forwardRef, useCallback, useEffect, useMemo, useRef } from "react";
 import { camelCase } from "lodash-es";
 import cx from "classnames";
 import { SIZES } from "../../constants";
@@ -15,6 +15,7 @@ import { ComponentDefaultTestId } from "../../tests/constants";
 import { backwardCompatibilityForProperties } from "../../helpers/backwardCompatibilityForProperties";
 import { getStyle } from "../../helpers/typesciptCssModulesHelper";
 import styles from "./Button.module.scss";
+import { useButtonLoading } from "./helper/useButtonLoading";
 
 export interface ButtonProps extends VibeComponentProps {
   children?: React.ReactNode;
@@ -143,7 +144,7 @@ const Button: VibeComponent<ButtonProps, unknown> & {
     },
     ref
   ) => {
-    const [loading, setLoading] = useState(isLoading);
+    const { loading } = useButtonLoading({ isLoading });
     const overrideDataTestId = backwardCompatibilityForProperties([dataTestId, backwardCompatabilityDataTestId]);
     const buttonRef = useRef<HTMLButtonElement>(null);
     useEffect(() => {
@@ -154,15 +155,6 @@ const Button: VibeComponent<ButtonProps, unknown> & {
       const buttonElement = buttonRef.current;
       buttonElement.style.color = getParentBackgroundColorNotTransparent(buttonElement, defaultTextColorOnPrimaryColor);
     }, [kind, buttonRef, color, defaultTextColorOnPrimaryColor]);
-
-    useEffect(() => {
-      const frameId = window.requestAnimationFrame(() => {
-        setLoading(isLoading);
-      });
-      return () => {
-        window.cancelAnimationFrame(frameId);
-      };
-    }, [isLoading]);
 
     const onMouseUp = useCallback(() => {
       const button = buttonRef.current;
